devtools/client/webconsole/new-console-output/test/components/test_message-repeat.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for MessageRepeat component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for MessageRepeat component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { MessageRepeat } = require("devtools/client/webconsole/new-console-output/components/message-repeat");

  const repeatRendered = renderComponent(MessageRepeat, { repeat: 99 });
  ok(repeatRendered.classList.contains("message-repeats"), "MessageRepeat has expected class");
  is(repeatRendered.style.visibility, "visible", "MessageRepeat with 2+ repeats is visible");
  is(repeatRendered.textContent, "99", "MessageRepeat shows correct number of repeats");

  const noRepeatRendered = renderComponent(MessageRepeat, { repeat: 1 });
  is(noRepeatRendered.style.visibility, "hidden", "MessageRepeat with 1 repeat is hidden");
  is(noRepeatRendered.textContent, "1", "MessageRepeat with 1 repeat shows correct number of repeats")

  SimpleTest.finish();
});
</script>
</body>
</html>
